+++ src/main/java/com/diplom/work/repo/RuleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RuleRepository extends JpaRepository<Rule,Integer> {
+public interface RuleRepository extends JpaRepository<Rule,Long> {
     List<Rule> findAllByOrderByClientNameAsc();
     Rule findByClientNumber(String clientNumber);
     List<Rule> findAllByOrderByIdAsc();
